@@ -1,11 +1,14 @@
 module restmock
 
-go 1.16
+go 1.19
 
 require (
 	github.com/gorilla/mux v1.8.0
-	go.uber.org/atomic v1.9.0 // indirect
-	go.uber.org/multierr v1.7.0 // indirect
 	go.uber.org/zap v1.18.1
 	gopkg.in/yaml.v2 v2.4.0
+)
+
+require (
+	go.uber.org/atomic v1.9.0 // indirect
+	go.uber.org/multierr v1.7.0 // indirect
 )
